+++ ocbpy/tests/test_boundary_eab.py
@@ -14,23 +14,6 @@ import ocbpy
 from . import test_boundary_ocb as test_ocb
 
 
-class TestEABoundaryDeprecations(test_ocb.TestOCBoundaryDeprecations):
-    """Test the deprecation warnings within the EABoundary class."""
-
-    def setUp(self):
-        """Initialize the test environment."""
-        self.test_class = ocbpy.EABoundary
-        test_dir = path.join(path.dirname(ocbpy.__file__), "tests",
-                             "test_data")
-        self.inst_init = {"instrument": "image", "hemisphere": 1,
-                          "filename": path.join(test_dir,
-                                                "test_north_circle")}
-
-    def tearDown(self):
-        """Clean up the test environment."""
-        del self.test_class, self.inst_init
-
-
 class TestEABoundaryLogFailure(test_ocb.TestOCBoundaryLogFailure):
     """Test the logging messages raised by the EABoundary class."""