+++ ocbpy/tests/test_cycle_boundary.py
@@ -50,26 +50,6 @@ class TestCycleMatchData(unittest.TestCase):
         del self.bad_time
         return
 
-    def test_deprecated_kwargs(self):
-        """Test warnings for deprecated kwargs in `get_next_good_ocb_ind`."""
-        # Set the deprecated keyword arguments with standard values
-        dep_inputs = {"min_sectors": 7, "rcent_dev": 8.0, "max_r": 23.0,
-                      "min_r": 10.0}
-        del_ind = self.rec_ind2 - self.rec_ind
-        test_times = np.arange(
-            self.ocb.dtime[self.ocb.rec_ind],
-            self.ocb.dtime[self.ocb.rec_ind + del_ind],
-            dt.timedelta(seconds=self.del_time)).astype(dt.datetime)
-
-        # Cycle through the keyword arguments that should raise a warning
-        for dkey in dep_inputs.keys():
-            kwargs = {dkey: dep_inputs[dkey]}
-            with self.subTest(kwargs=kwargs):
-                with self.assertWarnsRegex(DeprecationWarning,
-                                           "Deprecated kwarg will be removed"):
-                    self.test_func(self.ocb, test_times, idat=1, **kwargs)
-        return
-
     def test_bad_class_cycling_method(self):
         """Test raises ValueError when Boundary class missing cycling method."""